create RoutesWrap and put sidebar with modes there
